SA-1618: Moved variable to constants
